requirements.txt
openai
opencv-python
numpy
gradio
python-dotenv
